==> notebooks/1.2_WebScraping_Cities.ipynb <==
#Import required libraries
import pandas as pd
import requests
import re
from bs4 import BeautifulSoup
from geopy import Nominatim

# Scraping a list of world capitals

# Load thd page and get it into so yummy Beautiful Soup
html_doc = 'https://www.boldtuesday.com/pages/alphabetical-list-of-all-countries-and-capitals-shown-on-list-of-countries-poster'
page = requests.get(html_doc)
soup = BeautifulSoup(html_doc, 'html.parser')
soup = BeautifulSoup(page.content, 'html.parser')

print(soup)

data = []
table = soup.find('table')
table_body = table.find('tbody')

rows = table_body.find_all('tr')
for row in rows:
    cols = row.find_all('td')
    cols = [ele.text.strip() for ele in cols]
    data.append([ele for ele in cols if ele])


data

cities = []

for el in data[1:]:
    cities.append(el[1])

cities_df = pd.DataFrame(cities)

cities_df

cities_df.rename(columns = {0:'City'}, inplace = True)

len(cities_df)

cities_df.isnull().any()

# the length of dataframe is quite big, so I will randomly pick 100 cities
# divide into two:

sample = cities_df.sample(n = 100)

# Appending coordinates

# creating a geolocator to use it to get coordinates, increasing timeout.
geolocator = Nominatim(timeout = 5)

# applying the geolocator
sample['Latitude'] = sample['City'].apply(geolocator.geocode).apply(lambda x: x.latitude)

sample['Longitude'] = sample['City'].apply(geolocator.geocode).apply(lambda x: x.longitude)
