==> notebooks/2_Data_Cleaning.ipynb <==
import pandas as pd
import re
from geopy.geocoders import Nominatim

# Cleaning the 'races' file

races = pd.read_csv('../data/races.csv', index_col = 0)

races.head(10)

## Renaming columns

races.rename(columns = {'0':'Day', '1':'Month', '0.1':'Race', '1.1':'Circuit', '2':'Country'}, inplace = True)

races.head()

## Removing 'Tests'

For this project, I will focus only in scoring races which are called 'Grand Prix'.

# The new dataframe should not include races with 'Test' word in the 'Race' column.
races = races.loc[races['Race'].str.contains('Test') == False]

races.head()

## Updating dates

Now we removed the tests, all the dates are for same year 2020. I will create a new column with 'Date' as a DateTime

races['Date'] = pd.to_datetime((races['Day'].astype(str)+races['Month'] + '2020'), errors = 'coerce', infer_datetime_format=True)

races = races[['Date', 'Race', 'Circuit', 'Country']].reset_index(drop=True)

races.head()

races.dtypes

## Removing numbers on Race names

As they are already ordered, we can remove the numbers at the beginning of the race's name.

races['Race'] = races['Race'].apply(lambda x: re.sub(r'\d+ - ','',x))

races.head()

## Checking duplicates

Except for the country, the other columns should not have any duplicates.

races['Date'].duplicated(keep = False).any()

races['Race'].duplicated(keep = False).any()

races['Circuit'].duplicated(keep = False).any()

races.loc[races['Circuit'].duplicated(keep = False)]

The calendar from motogp.com has an error, and so the data scraped. As is just one mistake (circuit and country for Thailand Grand Prix) I will correct it manually

races.at[1,'Circuit'] = 'Buriram International Circuit'

races.at[1,'Country'] = 'THAILAND'

races.loc[races['Circuit'].duplicated(keep = False)]

## Looking for coordinates with GeoPy

# creating a geolocator to use it to get coordinates
geolocator = Nominatim()

# checking the geolocator
races['Circuit'].apply(geolocator.geocode)

As we can see above, the geolocator couldn't find the direction for row 4. Let's check the name.

races.loc[4, 'Circuit']

The problem may be due to the append 'Angel Nieto' to the name of the circuit. Let's remove it:

races.at[4, 'Circuit'] = 'Circuito de Jerez'

races.loc[4, 'Circuit']

Now I can try again to get the directions. But this time I will get directly the coordinates

races['Latitude'] = races['Circuit'].apply(geolocator.geocode).apply(lambda x: x.latitude)

races['Longitude'] = races['Circuit'].apply(geolocator.geocode).apply(lambda x: x.longitude)

races.head()

races.dtypes

import plotly.graph_objects as go

fig = go.Figure(data=go.Scattergeo(
        lon = races['Longitude'],
        lat = races['Latitude'],
        text = races['Circuit'],
        mode = 'markers',
        ))

fig.update_layout(
        title = 'MotoGP 2020 Circuits')

fig.show()

As seen in the map, the Phillip Island Circuit is not properly plotted. There is a mistake in its coordinates. 

PhillipIslandAddress = '381 Back Beach Rd, Ventnor VIC 3922'

races.query('Circuit == "Phillip Island"')

races.at[17,'Longitude'] = geolocator.geocode(PhillipIslandAddress).longitude

races.at[17,'Latitude'] = geolocator.geocode(PhillipIslandAddress).latitude

races.query('Circuit == "Phillip Island"')

fig = go.Figure(data=go.Scattergeo(
        lon = races['Longitude'],
        lat = races['Latitude'],
        text = races['Circuit'],
        mode = 'markers',
        ))

fig.update_layout(
        title = 'MotoGP 2020 Circuits')

fig.show()

## Races dataframe cleaned

races.to_csv('../data/races_cleaned.csv')

# to plot the circuits in Tableau Public, is better to have an excel file
races.to_excel('../data/races_cleaned.xlsx')

# Cleaning the 'riders' file

riders = pd.read_csv('../data/riders.csv', index_col = 0)

riders.head(10)

riders.isnull().sum()

## Renaming columns

riders.rename(columns = {'0':'Name', '1':'Team', '2':'Bike', '3':'BirthPlace'}, inplace = True)

riders.head()

## Removing 'Bike:' text in Bike column

The word 'Bike:' in the bike column is redundant information. Let's remove it.

riders['Bike'] = riders['Bike'].apply(lambda x: re.sub(r'Bike: ','',x))

riders.head()

## Checking duplicates

We can't have more than one rider with the same name.

riders['Name'].duplicated(keep = False).any()

## Checking errors in Teams and Bikes

As they can be repeated, let's check if all of them are written properly

riders['Team'].value_counts()

Although we have two different Teams for LCR Honda, due to main sponsors, both are the same team structure, so we will consider them as a unique Team

riders.loc[riders['Team'].str.contains('LCR')]

Another error has raised: Honda Team with KTM Bike is not possible. After fixing the Team, I will look for the same bike in Team and Bike

riders.at[8,'Team'] = 'LCR Honda'

riders.at[10,'Team']= 'LCR Honda'

riders['Team'].value_counts()

Checking if there is any other error between the name of the team and the bike. At least we have 1 in row[8]

riders.loc[riders.apply(lambda x: x['Bike'] in x['Team'], axis = 1) == False]

From the list above we can see that there are two Ducati Teams without the name of the bike in the name of the team: Reale Avintia Racing and Pramac Racing. But this is not an error.

Also, the Team Suzuki appears in the list due to capital letters, but this is not an error.

The unique error is the one it was raised before, having a KTM bike in the LCR Honda team, and this needs to be fixed.

riders.at[8,'Bike'] = 'Honda'

## Looking for coordinates with GeoPy

For this project I will consider the Birth Place as the riders' home. Although it may not be exactly the same place, is enough to get an approximate idea of the distance travelled during the whole season.

riders['Lat_Lon'] = riders['BirthPlace'].apply(geolocator.geocode).apply(lambda x: (x.latitude, x.longitude))

riders.head()

## Riders dataframe cleaned

riders.to_csv('../data/riders_cleaned.csv')
