==> notebooks/3.1_Analysis_Clustering.ipynb <==
import pandas as pd
from geopy import distance

# ANALYSIS

## The current calendar

# importing the DataFrame with the circuits, dates and coordinates
circuits = pd.read_csv('../data/races_cleaned.csv', index_col = 0)

circuits.head()

# Calculating the current distance. From one circuit to the previous one, and adding up
# each distance. This is a non-round distance:
current_distance = 0

for index in circuits.index[1:]:
    coord0 = (circuits.loc[index-1, 'Latitude'], circuits.loc[index-1, 'Longitude'])
    coord1 = (circuits.loc[index, 'Latitude'], circuits.loc[index, 'Longitude'])
    current_distance += round(distance.distance(coord0, coord1).km)

print('The current distance for the whole calendar is:', current_distance, 'km')

#### Current distance is more than 84k km. The objective is reduce this distance

## Clustering Circuits

I will use Unsupervised Learning algorithms to cluster the circuits, and then optimizing the distance between clusters.

### Clustering with DBSCAN
The first clustering will be using DBSCAN, as I don't want to force any number of clusters.

# first of all I need to create a list of lists with the coordinates:

coordinates = []
for index in circuits.index:
    coord_list = [circuits.loc[index,'Latitude'], circuits.loc[index,'Longitude']]
    coordinates.append(coord_list)

# As the latitude goes from -90º to 90º and the longitude from -180º to 180º, first of
# all I will scale the coordinates to give them the same weight.

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
coord_scale = scaler.fit_transform(coordinates)

coord_scale

from sklearn.cluster import DBSCAN

dbscan = DBSCAN(eps=0.5)
circuits_dbscan = dbscan.fit(coord_scale)
circuits['DBSCAN_Clusters'] = circuits_dbscan.labels_

circuits.groupby(['DBSCAN_Clusters']).size()

#### There is a cluster with 13 items and 7 outliers

circuits.groupby(['DBSCAN_Clusters', 'Country']).size()

#### The distribuiton makes sense, as the outliers are far from each other while the items in the cluster are pretty close.

#### For further analysis, I will split them into two new dataframes according to the first clustering.

outliers = circuits.query('DBSCAN_Clusters == -1').reset_index(drop = True)

outliers

big_cluster = circuits.query('DBSCAN_Clusters == 0').reset_index(drop = True)

big_cluster

### Clustering with KMeans
In order to small calculations of distances, I will split the cluster with 13 items into several clusters of max 5 items each. This will simplify the optimizing calculation for the routes between the circuits into the cluster.

# creating the coordinates for circuits in big_cluster

cluster_coord = []
for index in big_cluster.index:
    coord_list = [big_cluster.loc[index,'Latitude'], big_cluster.loc[index,'Longitude']]
    cluster_coord.append(coord_list)

cluster_coord

# after different attemps, to get a max of 5 items per cluster, I need 4 clusters
from sklearn.cluster import KMeans

kmeans = KMeans(n_clusters=4)
small_clusters = kmeans.fit(cluster_coord)
small_clusters.labels_

# creating a new column to write the code of subcluster
big_cluster['Subcluster'] = small_clusters.labels_

big_cluster

big_cluster['Subcluster'].value_counts()

#### As seen in the subclusters, there is a group with only 1 circuit. That means that in fact, we can consider this circuit as an outlier as well. I will drop this row from the big_cluster and append it into the otuliers dataframe.

outliers = outliers.append(big_cluster.query('Subcluster == 3')\
                           [outliers.columns]).reset_index(drop = True)

outliers

big_cluster.drop(big_cluster.query('Subcluster == 3').index,inplace = True)

big_cluster.reset_index(drop = True)

### In order to not reclustering each time I run the code, I will save the DF in csv files and upload them from a new notebook.

outliers.to_csv('../data/outliers.csv')

big_cluster.to_csv('../data/big_cluster.csv')

outliers.to_excel('../data/outliers.xlsx') # needed for Tableau Public

big_cluster.to_excel('../data/big_cluster.xlsx') # needed for Tableau Public

complete = big_cluster.append(outliers, sort=False)

complete = complete.fillna(value = 99) # to have a complete different number on outliers

complete.to_excel('../data/clusters_df.xlsx')
