==> notebooks/3.3_Analysis_Optimizing.ipynb <==
import pandas as pd
from geopy import distance
import numpy as np

# ANALYSIS

# importing the cluster's dataframes
outliers = pd.read_csv('../data/outliers.csv', index_col = 0)
big_cluster = pd.read_csv('../data/big_cluster.csv', index_col = 0)

outliers.head()

big_cluster.head()

big_cluster['Subcluster'].value_counts()

#### I will start optimizing internal routes for subclusters.

## Optimizing routes

As the result of applying mlrose is in 'units' and the units of our coordinates are degrees of Latitude and Longitude, it is better to apply mlrose with a matrix of distances previously calculated in km with geopy.

subclusters = np.sort(big_cluster['Subcluster'].unique())

np.sort(subclusters)

# creating new dataframes for subclusters, so after we can retrieve information
sub0 = big_cluster.loc[big_cluster['Subcluster'] == subclusters[0]].reset_index(drop=True)
sub1 = big_cluster.loc[big_cluster['Subcluster'] == subclusters[1]].reset_index(drop=True)
sub2 = big_cluster.loc[big_cluster['Subcluster'] == subclusters[2]].reset_index(drop=True)

from functions import create_dist_list

# creating dist lists for all subclusters, applying the function 'create_dist_list':
subcluster_dist_dict = {}
subcluster_dist_dict[0] = create_dist_list(sub0)
subcluster_dist_dict[1] = create_dist_list(sub1)
subcluster_dist_dict[2] = create_dist_list(sub2)

subcluster_dist_dict

# Calculating best routes:
from functions import best_route

routes_dict = {}

for cluster in subclusters:
    best_order, best_dist = best_route(subcluster_dist_dict[cluster])
    routes_dict[cluster] = (best_order, best_dist)

routes_dict

### At this point, I have the best routes for the circuits inside the 3 subclusters. Now I will calculate the best route for the outliers

outliers_dist_list = create_dist_list(outliers)

outliers_dist_list

outliers_best_route = best_route(outliers_dist_list)

### At this point, I have all the distances between points inside each cluster, considering 'the outliers' like a cluster, but these are round distances.

### Now it's time to make them linear and calculate distances from one cluster to another and sum them all up.

### To do this, as I have one circuit in the 'outliers' which initally was considered in the first calculated cluster, I will consider this circuit as the last stage of the first route. And from this circuit, I will calculate the distance to the nearest circuit of the 3 subclusters.

### I will follow this method until I have all the circuits connected.

## STAGE 1

# looking for the index of the wanted last circuit, in outliers
s1_last = outliers.query('DBSCAN_Clusters == 0').index[0]

# importing needed functions
from functions import reordering_last
from functions import reordering_first
from functions import to_subtract

# applying functions:
s1 = reordering_last(outliers_best_route[0], s1_last)
s1_sub = to_subtract(s1, outliers_dist_list)

# calculating the distance considering s1 an open route:
s1_dist = outliers_best_route[1] - s1_sub

# creating the list for the whole calendar and the total length:
new_calendar = []
new_length = 0

# appending names for this stage
for i in s1:
    name = outliers.loc[i, 'Circuit']
    new_calendar.append(name)

# adding stage length
new_length += round(s1_dist)

# calculating the shortest distance from last point of stage 1 to the others
# this will determine the cluster for second stage

checking_dist = {}

for idx in big_cluster.index:
    start_lat = outliers.loc[s1_last,'Latitude']
    start_lon = outliers.loc[s1_last,'Longitude']
    start_coord = (start_lat, start_lon)
    
    i_lat = big_cluster.loc[idx,'Latitude']
    i_lon = big_cluster.loc[idx,'Longitude']
    i_coord = (i_lat, i_lon)
    
    dist = distance.distance(start_coord, i_coord).km
    if dist > 0:
        checking_dist[dist] = big_cluster.loc[idx,'Circuit']

# assigning the distance from stage 1 to stage 2 = min distance from last point of
# cluster 'outliers' and first point of next cluster.
s1_s2_dist = min(checking_dist.keys())

# updating new_lenght
new_length += round(s1_s2_dist)

## STAGE 2

# checking the name of the circuit for the next cluster
s2_first_name = checking_dist[s1_s2_dist]

# getting the index of the next cluster from circuit name
cluster_s2 = big_cluster.loc[big_cluster['Circuit'] == s2_first_name]['Subcluster'].values[0]

print(cluster_s2)

# getting the index of the subcluster for the first circuit of second cluster
s2_first = sub1.loc[sub1['Circuit'] == s2_first_name].index[0]

# applying functions to reorder and get distance to subtract:
s2 = reordering_first(routes_dict[cluster_s2][0], s2_first)
s2_sub = to_subtract(s2, subcluster_dist_dict[cluster_s2])

# calculating the distance considering s2 an open route:
s2_dist = routes_dict[cluster_s2][1] - s2_sub

# appending names for this stage
for i in s2:
    name = sub1.loc[i, 'Circuit']
    new_calendar.append(name)

# adding stage length
new_length += round(s2_dist)

# calculating the shortest distance from last point of stage 2 to the others
# this will determine the cluster for third stage

checking_dist = {}

for idx in big_cluster.loc[big_cluster['Subcluster'] != 1].index:
    start_lat = sub1.loc[s2[-1],'Latitude']
    start_lon = sub1.loc[s2[-1],'Longitude']
    start_coord = (start_lat, start_lon)
    
    i_lat = big_cluster.loc[idx,'Latitude']
    i_lon = big_cluster.loc[idx,'Longitude']
    i_coord = (i_lat, i_lon)
    
    dist = distance.distance(start_coord, i_coord).km
    if dist > 0:
        checking_dist[dist] = big_cluster.loc[idx,'Circuit']

# assigning the distance from stage 2 to stage 3 = min distance from last point of
# cluster 'sub1' and first point of next cluster.
s2_s3_dist = min(checking_dist.keys())

# updating new_lenght
new_length += round(s2_s3_dist)

## STAGE 3

# checking the name of the circuit for the next cluster
s3_first_name = checking_dist[s2_s3_dist]

# getting the index of the next cluster from circuit name
cluster_s3 = big_cluster.loc[big_cluster['Circuit'] == s3_first_name]['Subcluster'].values[0]

print(cluster_s3)

# getting the index of the subcluster for the first circuit of second cluster
s3_first = sub2.loc[sub2['Circuit'] == s3_first_name].index[0]

# applying functions to reorder and get distance to subtract:
s3 = reordering_first(routes_dict[cluster_s3][0], s3_first)
s3_sub = to_subtract(s3, subcluster_dist_dict[cluster_s3])

# calculating the distance considering s3 an open route:
s3_dist = routes_dict[cluster_s3][1] - s3_sub

# appending names for this stage
for i in s3:
    name = sub2.loc[i, 'Circuit']
    new_calendar.append(name)

# adding stage length
new_length += round(s3_dist)

# calculating the shortest distance from last point of stage 3 to the last cluster

checking_dist = {}

for idx in big_cluster.loc[big_cluster['Subcluster'] == 0].index:
    start_lat = sub2.loc[s2[-1],'Latitude']
    start_lon = sub2.loc[s2[-1],'Longitude']
    start_coord = (start_lat, start_lon)
    
    i_lat = big_cluster.loc[idx,'Latitude']
    i_lon = big_cluster.loc[idx,'Longitude']
    i_coord = (i_lat, i_lon)
    
    dist = distance.distance(start_coord, i_coord).km
    if dist > 0:
        checking_dist[dist] = big_cluster.loc[idx,'Circuit']

# assigning the distance from stage 3 to stage 4 = min distance from last point of
# cluster 'sub2' and first point of last cluster.
s3_s4_dist = min(checking_dist.keys())

# updating new_lenght
new_length += round(s3_s4_dist)

## STAGE 4

# checking the name of the circuit for the last cluster
s4_first_name = checking_dist[s3_s4_dist]

# getting the index of the next cluster from circuit name
cluster_s4 = big_cluster.loc[big_cluster['Circuit'] == s4_first_name]['Subcluster'].values[0]

print(cluster_s4) # should be 0!

# getting the index of the subcluster for the first circuit of second cluster
s4_first = sub0.loc[sub0['Circuit'] == s4_first_name].index[0]

# applying functions to reorder and get distance to subtract:
s4 = reordering_first(routes_dict[cluster_s4][0], s4_first)
s4_sub = to_subtract(s4, subcluster_dist_dict[cluster_s4])

# calculating the distance considering s3 an open route:
s4_dist = routes_dict[cluster_s4][1] - s4_sub

# appending names for this stage
for i in s4:
    name = sub0.loc[i, 'Circuit']
    new_calendar.append(name)

# adding stage length
new_length += round(s4_dist)

## LAST CHECKS

# the length of new_calendar should be = 20
len(new_calendar) == 20

# the values of new_calendar should be all unique
len(new_calendar) == len(set(new_calendar))

# the new distance should be lower than the first calculated: 84594 km
current_length = 84594
new_length < current_length

# FINAL RESULTS

## NEW CALENDAR

print('The optimized calendar should be:')
for item in new_calendar:
    print(item)

current_calendar = pd.read_csv('../data/races_cleaned.csv')

# df.reindex(["Z", "C", "A"])
big_cluster_dropped = big_cluster.drop(columns = 'Subcluster')
new_df = outliers.append(big_cluster_dropped)
new_df = new_df.set_index('Circuit')
new_df = new_df.reindex(new_calendar)
new_df = new_df.reset_index()
new_df.drop(columns = ['Date', 'DBSCAN_Clusters'], inplace = True)
new_df['Date'] = current_calendar['Date'] # keeping the same dates, but new order of circuits

new_df

new_df.to_excel('../data/new_calendar.xlsx')

## NEW DISTANCE

print('The total distance for the new calendar is:', new_length)
print('That means a reduction of:',round(100*(current_length-new_length)/current_length, 1),'%')

# CONCLUSIONS

There are positive and negative conclusions about these results. 

Negative conclusions:

- Clustering circuits with MLRose algorithm offers the best round routes for the cluster itself, but is not useful enough if we want to get the whole minimum distance.
- The distance calculated is taking into account that we travel from one circuit to the next, without considering that the riders and staff may go home between races.

Positive conclusions:

- The new calendar reduced the travelled distance by almost 40%
- These route could be followed by trucks transporting bikes and material
- There is room for improvement
